refactor: discard unneeded events from bpf (#3)
This commit refactors the way functions are acknowledged from
kernel space. Only new functions that are not acknowledged yet
are sent to user space via the BPF ring buffer.

This dramatically improved performance and reduces resource utilization,
moving close to zero possibility of dropped events due to the
ring buffer being fulfilled because of user space not consuming fast
enough.

Signed-off-by: Massimiliano Giovagnoli <maxgio92@pm.me>