feat: add minimal healthcheck server
A minimal healthchceck is introduced with this commit.

An UDS socket is created within tracer.Init, and a readiness
message is sent over it as 0x01 indicating that the tracer is
ready being that it is tracing all expected functions.
Readiness here means that userspace can execute all the tests
because xcover tracer will capture them all.

Signed-off-by: Massimiliano Giovagnoli <maxgio92@pm.me>