Daemonize profiler and refactoring (#17)
* feat(cmd): daemonize profiling and refactoring

This commit enable the daemonization of the profiler, and introduces
new command to manage the profiling as a daemon.
`start`, `stop`, `status` commands are added to control the profiling
when the profiling is `run` in detached mode, via the `--detach` flag.

Furthermore, it moves the BPF logics related to the tracer probe to a dedicated
package. Trace package should be as much as possible agnostic
to the technology leveraged to trace userland functions.

Signed-off-by: Massimiliano Giovagnoli <maxgio92@pm.me>